+++ src/js/main/design.js
@@ -4,17 +4,26 @@ import DOM from "../ui/DOM/Classes/dom"
 import { $$ } from "../services/Language/handler"
 import { SettingsSectionElement, SettingsGroupContainer } from "../ui/DOM/Library/settings"
 import { RadioLabel } from "../ui/DOM/Library/object/input"
+import TextWithSubtext from "../ui/DOM/Library/object/textWithSubtext"
+import SettingsStorage from "../services/Settings/SettingsStorage"
+import reloadToast from "../tools/interaction/reloadToast"
 
 export default class Design {
     static theme = "default"
 
     static defaultTheme = "default"
 
+    static presetDarkTheme = "dark"
+
     static themeObject = false
 
     static LSName = "temply_ui_theme"
 
     static themesList = themesList
+
+    static basedOnStack = []
+
+    static isAutomatic = false
 
     static getVar(name, value = false, float = false) {
         if (!value) return `var(--${name})`
@@ -27,15 +36,46 @@ export default class Design {
         return get
     }
 
-    static async themeLoader(name = null) {
+    static async dafaultThemeHandler() {
+        const loadDark = async () => {
+            await this.themeLoader(this.presetDarkTheme, true)
+            this.isAutomatic = true
+        }
+
+        const loadDefault = async () => {
+            await this.themeLoader(this.defaultTheme, true)
+            this.isAutomatic = true
+        }
+
+        const DarkMediaQuery = window.matchMedia("(prefers-color-scheme: dark)")
+        const isDark = DarkMediaQuery.matches
+
+        if (!this.isAutomatic) {
+            DarkMediaQuery.addListener((ev) => {
+                if (!this.isAutomatic) return
+                if (ev.matches) loadDark()
+                else loadDefault()
+            })
+        }
+
+        this.isAutomatic = true
+
+        return (isDark ? this.presetDarkTheme : this.defaultTheme)
+    }
+
+    static async themeLoader(name = null, auto = false) {
         let newTheme = false
+        auto = auto || false
 
         // 1. Get requested theme name
         name = name || localStorage.getItem(this.LSName)
-        if (name === null) return false
+        if (name === null) {
+            name = await this.dafaultThemeHandler()
+            auto = true
+        }
 
         // 2. Get theme object if not default
-        if (name !== this.defaultTheme) {
+        if (name !== this.defaultTheme && name !== this.theme) {
             try {
                 newTheme = await require(`Resources/styles/colors/${name}/theme.css`)
             } catch (e) {
@@ -46,7 +86,7 @@ export default class Design {
 
         // 3. Disconnect current theme if present
         try {
-            if (this.themeObject) {
+            if (this.themeObject && name !== this.theme) {
                 this.themeObject.unuse()
             }
         } catch (e) {
@@ -77,9 +117,10 @@ export default class Design {
         }
 
         // 6. Apply new settings
-        this.themeObject = newTheme
+        if (newTheme) this.themeObject = newTheme
         this.theme = name
-        localStorage.setItem(this.LSName, name)
+        if (!auto) localStorage.setItem(this.LSName, name)
+        if (!auto) this.isAutomatic = false
 
         return this.themeObject
     }
@@ -103,7 +144,7 @@ export default class Design {
         return str.match(/^rgba?\((\d+),\s*(\d+),\s*(\d+)(?:,\s*(\d+(?:\.\d+)?))?\)$/).slice(1)
     }
 
-    static generateThemesList(act) {
+    static async generateThemesList(act) {
         const self = this
         const themes = this.themesList
 
@@ -140,13 +181,97 @@ export default class Design {
                 if (!s) return
                 self.themeLoader(theme.dir)
             },
-            selected: () => (theme.dir === this.theme),
+            selected: () => (theme.dir === this.theme && self.isAutomatic !== true),
         }))
+
+        list.unshift({
+            content: new TextWithSubtext(
+                $$("@settings/appearance/themes/names/system"),
+                $$("@settings/appearance/themes/system_theme_info"),
+            ),
+            handler(s) {
+                if (!s) return
+                localStorage.removeItem(self.LSName)
+                self.themeLoader()
+            },
+            selected: () => (self.isAutomatic === true),
+        })
 
         act.createSection({ id: "theme-selection-section", dom: SettingsSectionElement, options: { name: $$("@settings/appearance/themes") } })
             .getSection("theme-selection-section")
             .createGroup({ id: "theme-selection-group", dom: SettingsGroupContainer, options: {} })
             .getGroup("theme-selection-group")
             .createItem({ id: "theme-selection-chooser-radios", dom: ({ items }) => new DOM({ new: "div", content: new RadioLabel(items, [], true) }), options: { items: list } })
+
+
+        const AWT = !!(await SettingsStorage.getFlag("ui_wm_adv_transitions"))
+        const AWTCSS = !!(await SettingsStorage.getFlag("ui_wm_adv_css_transitions"))
+        const NoAnimation = !!(await SettingsStorage.getFlag("ui_wm_no_transitions"))
+
+        const windowTransitions = [
+            {
+                content: new TextWithSubtext(
+                    $$("@settings/appearance/animations/optimized"),
+                    $$("@settings/appearance/animations/optimized/info"),
+                ),
+                handler(s) {
+                    if (!s) return
+                    SettingsStorage.setFlag("ui_wm_no_transitions", 0)
+                    SettingsStorage.setFlag("ui_wm_adv_transitions", 1)
+                    SettingsStorage.setFlag("ui_wm_adv_css_transitions", 1)
+                    reloadToast()
+                },
+                selected: () => (AWT && AWTCSS && !NoAnimation),
+            },
+            {
+                content: new TextWithSubtext(
+                    $$("@settings/appearance/animations/simple"),
+                    $$("@settings/appearance/animations/simple/info"),
+                ),
+                handler(s) {
+                    if (!s) return
+                    SettingsStorage.setFlag("ui_wm_no_transitions", 0)
+                    SettingsStorage.setFlag("ui_wm_adv_transitions", 0)
+                    SettingsStorage.setFlag("ui_wm_adv_css_transitions", 0)
+                    reloadToast()
+                },
+                selected: () => (!AWT && !NoAnimation),
+            },
+            {
+                content: new TextWithSubtext(
+                    $$("@settings/appearance/animations/stable"),
+                    $$("@settings/appearance/animations/stable/info"),
+                ),
+                handler(s) {
+                    if (!s) return
+                    SettingsStorage.setFlag("ui_wm_no_transitions", 0)
+                    SettingsStorage.setFlag("ui_wm_adv_transitions", 1)
+                    SettingsStorage.setFlag("ui_wm_adv_css_transitions", 0)
+                    reloadToast()
+                },
+                selected: () => (AWT && !AWTCSS && !NoAnimation),
+            },
+            {
+                content: new TextWithSubtext(
+                    $$("@settings/appearance/animations/no_animation"),
+                    $$("@settings/appearance/animations/no_animation/info"),
+                ),
+                handler(s) {
+                    if (!s) return
+                    SettingsStorage.setFlag("ui_wm_no_transitions", 1)
+                    SettingsStorage.setFlag("ui_wm_adv_transitions", 0)
+                    SettingsStorage.setFlag("ui_wm_adv_css_transitions", 0)
+                    reloadToast()
+                },
+                selected: () => (!AWT && !AWTCSS && NoAnimation),
+            },
+
+        ]
+
+        act.createSection({ id: "window-transition-section", dom: SettingsSectionElement, options: { name: $$("@settings/appearance/window_transition") } })
+            .getSection("window-transition-section")
+            .createGroup({ id: "window-transition-group", dom: SettingsGroupContainer, options: {} })
+            .getGroup("window-transition-group")
+            .createItem({ id: "window-transition-radios", dom: ({ items }) => new DOM({ new: "div", content: new RadioLabel(items, [], true) }), options: { items: windowTransitions } })
     }
 }